debug trivial case

+++ src/msolve/msolve-data.h
@@ -199,7 +199,7 @@ typedef msolvetrace_data_struct mstrace_t[1];
 
 typedef struct{
   int dim;
-  int dquot;
+  long dquot;
   int32_t ht_size; /* initial_hts */
   int32_t nr_threads;
   int32_t max_nr_pairs;
